changed floats to doubles

+++ src/pl/sda/Main.java
@@ -6,15 +6,15 @@ public class Main {
     public static void main(String[] args) {
         Scanner input = new Scanner(System.in);
         System.out.println("What is your weight in kg");
-        float weight = input.nextFloat();
+        double weight = input.nextDouble();
         System.out.println("What is your height in m");
-        float height = input.nextFloat();
-        float bmi = bmi(weight, height);
+        double height = input.nextDouble();
+        double bmi = bmi(weight, height);
         System.out.println("Your BMI is: " + bmi);
 
     }
 
-    private static float bmi(float weight, float height) {
+    private static double bmi(double weight, double height) {
         return weight/(height * height);
     }
 }
